refactor: rename `erros.js` to `errors.js`
