--- infra/erros.js
@@ -1,54 +1,0 @@
-export class InternalServerError extends Error {
-  constructor({ cause, statusCode }) {
-    super("An unexpected internal error occurred.", {
-      cause,
-    });
-    this.name = "InternalServerError";
-    this.action = "Please try again later or contact support.";
-    this.statusCode = statusCode || 500;
-  }
-  toJSON() {
-    return {
-      name: this.name,
-      message: this.message,
-      action: this.action,
-      status_code: this.statusCode,
-    };
-  }
-}
-
-export class ServiceError extends Error {
-  constructor({ cause, message }) {
-    super(message || "Services currently unavailable.", {
-      cause,
-    });
-    this.name = "ServiceError";
-    this.action = "Check if the service is available.";
-    this.statusCode = 503;
-  }
-  toJSON() {
-    return {
-      name: this.name,
-      message: this.message,
-      action: this.action,
-      status_code: this.statusCode,
-    };
-  }
-}
-
-export class MethodNotAllowedError extends Error {
-  constructor() {
-    super("Method not allowed for this endpoint.");
-    this.name = "MethodNotAllowedError";
-    this.action = "Contact Support.";
-    this.statusCode = 405;
-  }
-  toJSON() {
-    return {
-      name: this.name,
-      message: this.message,
-      action: this.action,
-      status_code: this.statusCode,
-    };
-  }
-}
