+++ models/migrator.js
@@ -1,7 +1,7 @@
 import migrationRunner from "node-pg-migrate";
 import { resolve } from "node:path";
 import database from "infra/database.js";
-import { ServiceError } from "infra/erros.js";
+import { ServiceError } from "infra/errors.js";
 
 const defaultMigrationOptions = {
   dryRun: true,
